Move `span_map` file to the right folder
