+++ src/librustdoc/html/span_map.rs
